+++ src/validators/exceptions/ValidatorNotFoundException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace xamned\framework\validators\exceptions;
+
+class ValidatorNotFoundException extends \Exception
+{
+    // ...
+}